Load clients before syncing one.
A common pattern is to write a new config, and then hit sync for it. So we
need to make sure we've got all clients loaded, or else we might just reply
with a 404.

We could refactor loading clients to have a method to only load the one we're
interested in, but since it's cheap to do, we can just load them all.