Rename to "package keysync", move main.go to cmd subpackage
This allows easier use of the Keysync code as a library, which might be
important if you want to integrate code not suitable for the open source
library, such as a custom logrus hook, or simply too opinionated for a shared
open source project.

We debated in person a bit about whether to do it this way, or move non-main
code into a `lib` folder.  I think this way makes it easier to work with the
codebase, since it's `package keysync` instead of lib.  Having a `cmd` folder
also allows us to add more commands in the future easily.