Configure a base secrets directory, and subdirectories for each client
This makes keysync easier to configure, because it's one less thing per-client
In practise for us, the key for each client entry in the yaml is what we want
that directory to be called, so this means we can simplify our configs.  It
also gives us an easier migration path because two syncers can share a client
config folder (which is host-specific), and sync to different locations based
on a different global configuration.