Make the fallback logic for uid/gid of files independent of each other
Before, if one of the user or group wasn't found, both the user and group used
the fallback.  That isn't what was desired.

This also removes use of two global variables for the passwdFile and groupFile.

I've reorganized the tests a bit to match the new structure of the code.