Fix bug around empty vs unspecified usernames
Because the empty username is invalid, and it's what we get from the yaml if
a string is unspecified, we get a bit of confusion around which case occurs.

This results in an error getting logged each time we have an unspecified user
or group, which is annoying, and could potentially cause problems if you did
have an /etc/passwd entry with a blank username.