Remove entries from oldClients after cleanup
This fixes issue #45

Because we didn't remove them from this list, we'd try to clean up after each
sync, even if the client had been recreated.

This is a bit annoying to test, so I'm going to refactor syncer and our tests
a bit to make this class of bug testable, however this is causing some problems
in production so I'd like to get this fix out before the weekend.