Return early if there's no SyncState for a secret
Add a bit more logging to when FileInfo changes so we can see what the
difference is.  Explicitly name the first field to avoid embedding.

Since the checksum wouldn't match the zero value, this wouldn't cause any
problems, except that some old test data shipped with Keywhiz has a zero
checksum, so this printed some stray messages.  While that shouldn't happen in
a real server, this is generally less likely to cause trouble.