Refactor all filesystem interaction out of syncer
This adds two new interfaces (OutputCollection and Output) which are used to
split out all the code for writing to disk.  This means the syncing logic is
much smaller and easier to read and test.

There's a new InMemoryOutput that implements the same interface for use in
testing the syncer seperately from writing output.  This allows us to remove
test cases that had to exercise the syncer to test file writing, and have
those tests operate seperately.