Use exponential backoff between retries
MMMMretries, min backoff time and max backoff can be specified in main
keysync config with overrides available for each client. The config
takes a string and later attemps to parse the duration using
`time.ParseDuration` - this behavior is analogous to `poll_interval`.

As part of the refactor, the timeout value is now configurable (defaults
to 60s in absence of keysync's config) and also can be overridden by
each client.