Check that secret's filename override is not a filepath
Keysync assumes that any Keywhiz 404 response to a GET /secret means
that a secret has been deleted and it attempts to delete that secret
from a local filesystem. The operational assumption is that a secret
name's cannot be a filepath. Keysync properly checks for that before
writing a file to disk, but not before attempting to delete a stale
secret. This opens up Keysync to a vulnerability that would potentially
allow to delete any file on the system. This change addresses the issue.