More robust sync mechanism (#63)
- Document current state of managing dependencies
- Add an option to retry failed requests to the server
  * Retries can be specified in the config with `max_retries`. Defaults to 1.
- Use exponential backoff between retries

Max retries, min backoff time and max backoff are be specified in main
keysync config. The config takes a string and later attempts to parse the duration 
using `time.ParseDuration` - this behavior is analogous to `poll_interval`.

As part of the refactor, the timeout value is now configurable (defaults
to 60s in absence of keysync's config).