Delete secrets when syncing a removed client
Previously, we'd wait until the next full sync, which was a bit confusing, and
a bit hard to test.

This removes the somewhat awkward "oldClients" state from the Syncer and just
returns the clients that need to be cleaned up.  That simplifies the RunOnce
function, which was a bit big too.

Extends the integration test to check client deletion works.