Implement Keysync Backup/Restore
This is a first pass at implementing an encrypted backup feature for keysync.
This writes an AES-GCM encrypted tar file to a configured disk location,
containing the secrets keysync has written out to the regular secrets location.

The Keyrestore program is updated to take these encrypted files and restore
them.

This is useful if you want to avoid writing plaintext secrets to disk, but want
a way to reboot or recover servers without Keywhiz running (eg, in an outage
scenario).

Right now a fixed AES key is used.  It is loaded after syncing completes, so
you can use a key stored in Keywhiz.  Recovering that key is left as an
exercise to the reader, or future improvements to this tool.