Refactor file writing out to decouple from Secrets
This makes a new function which handles atomically writing files with a given
mode and permissions.  I plan to reuse this code in the future, so this is just
the first part of seperation

The package is called "output" because I plan to move the Output* code there
too, but that requires a bit more refactoring to avoid circular dependencies.
So that's another PR for later.

Normalize uid and gid to int, instead of uint32, as int is what f.Chown takes.