Rewrite Keysync syncer to use new endpoint to retrieve batched secrets
Keywhiz previously exposed two endpoints: one which listed the
names and metadata, but not contents, for the secrets associated
with an app, and one which returned the contents for a single
requested secret. Keywhiz recently added a third endpoint, which
returns the contents (and other metadata) for a list of
requested secrets.

This updates the syncer to attempt to use this new endpoint to
retrieve secret contents, falling back to the single-secret
endpoint if any of the requested secrets are missing. In the
common case where secrets are not deleted between when Keysync
lists secrets and when it retrieves the contents for any
missing/updated secrets, this will result in Keysync making only
one (slower) request to Keywhiz to sync new/changed secrets,
rather than one request to Keywhiz per new/changed secret.